fix implicite time_t -> uint cast in Uuid
Former-commit-id: 377d89365166b00ed96dca9dbcb8f5d9f903ac7d

+++ framework/include/minko/Uuid.hpp
@@ -43,7 +43,7 @@ namespace minko
 		{
 			if (!_randSeeded)
 			{
-				std::srand(std::time(0));
+				std::srand(static_cast<uint>(std::time(0)));
 				_randSeeded = true;
 			}
